pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "asyncpg>=0.30.0",
    "ffmpeg-python>=0.2.0",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "tlgbotfwk>=0.4.61",
]
